crud +search+ map + affichage par date+qrcode

+++ src/main/java/com/fogbank/springsecurity/controller/LostProductController.java
@@ -20,7 +20,7 @@ public class LostProductController {
     }
     @PostMapping("/listeLostP/{idp}/{idu}")
     @ResponseBody
-    void ajouterEtaffecterListeLostProduct(@PathVariable("idp") Integer idp,@PathVariable("idu") Long idu) {
+    void ajouterEtaffecterListeLostProduct(@PathVariable("idp") Integer idp,@PathVariable("idu") Integer idu) {
         lostProductService.ajouterEtaffecterListeLostProduct(idp, idu);
     }
     @GetMapping
