+++ src/main/java/com/fogbank/springsecurity/services/ProductService.java
@@ -11,9 +11,9 @@ public interface ProductService {
     Product updateProduct(Integer id,Product product);
     List<Product> getProductsByUserId(Integer userId);
     Product findProduct(Integer id);
-    Product addProductU(String type, String picture, String description, String name, String phone, String email, String address, Long userId);
+    Product addProductU(String type, String picture, String description, String name, String phone, String email, String address, Integer userId);
 
-    void ajouterEtaffecterListeProduct (List<Product> lb, Long iduser);
+    void ajouterEtaffecterListeProduct (List<Product> lb, Integer iduser);
     Product getProductById(Integer id);
 
 }
